Split use case with parent logic
